refactor: Extract game result logic into useGameResult hook
